LayoutTests/http/tests/media/video-query-url-expected.txt


Tests that CyberKit is able to open a media file specified with a query url.
EVENT(canplay)
END OF TEST
